Minor mistake in util

+++ exhaustive/util.c
@@ -13,7 +13,7 @@ void printU8 (uint8_t x){
   printf("\n");
 }
 
-void printU64(uint8_t x){
+void printU64(uint64_t x){
   for(int i = 0; i < 64; i++){
     printf((x & (1l << i))? "#" : " ");
   }
