+++ exhaustive/util.h
@@ -12,7 +12,7 @@
 
 
 void printU8 (uint8_t);
-void printU64(uint8_t);
+void printU64(uint64_t);
